# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OCMS & Store LaPet Omnichannel Master API
  description: >
    Hợp đồng giao tiếp API chuẩn hóa toàn diện (Master OpenAPI Specification) giữa
    Hệ thống Quản trị Vận hành Đa kênh OCMS và Website Bán Hàng / Cổng Đại lý (store-lapet),
    đồng thời tích hợp liên thông với ERP Odoo và Kênh Zalo OA.
    
    Tài liệu này bao quát toàn bộ các nhóm tính năng vận hành:
    1. **Orders (Đơn hàng & Tiếp nhận chuẩn):** Cổng tiếp nhận đơn hàng bất đồng bộ, tra cứu trạng thái, duyệt đơn Odoo, hủy đơn và thống kê.
    2. **Products & Inventory (Sản phẩm & Tồn kho):** Danh mục sản phẩm, bảng giá đại lý/bán lẻ, kiểm tra tồn kho Odoo thời gian thực.
    3. **Promotions & Pricing Engine (Khuyến mãi & Động cơ tính giá):** Động cơ tính chiết khấu bậc thang tự động cho giỏ hàng, tư vấn ưu đãi bằng AI.
    4. **Customers & Dealers (Khách hàng & Đại lý B2B):** Xác thực OTP/Mật khẩu, hồ sơ đại lý, hạn mức công nợ, doanh số và quản lý danh bạ CRM.
    5. **Omnichannel Chat (Hội thoại & Zalo):** Quản lý luồng trò chuyện, gửi tin nhắn trao đổi, tự động bóc tách đơn hàng AI từ chat.
    6. **Webhooks & Sync (Thanh toán & Đồng bộ):** Webhook thanh toán tự động VietQR và kích hoạt đồng bộ dữ liệu Odoo ERP.
  version: 2.0.0
  contact:
    name: OCMS Integration Architecture Team
    email: dev@lapet.vn
servers:
  - url: http://localhost:3000
    description: Môi trường phát triển cục bộ (Local Development)
  - url: https://api.ocms.lapet.vn
...
paths:
...
  /api/v1/router/odoo/update-order:
    post:
      tags:
        - Universal Router
      summary: Router Gateway - Điều phối cập nhật đơn hàng sang máy chủ Odoo đích (Test/Prod)
      description: >
        Nhận gói cập nhật chi tiết đơn hàng (dòng sản phẩm, số lượng, đơn giá, chiết khấu, ghi chú),
        Router tự động chuyển tiếp sang máy chủ Odoo đích (odoo_test hoặc odoo_prod) theo cấu hình active.
      operationId: routerUpdateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouterUpdateOrderPayload'
      responses:
        '200':
          description: Cập nhật đơn hàng thành công qua Router Gateway
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  odoo_order_id:
                    type: integer
                    example: 2374
                  updated:
                    type: boolean
                    example: true
                  target:
                    type: string
                    example: "odoo_test"
                  target_name:
                    type: string
                    example: "Odoo ERP (Môi trường Test)"
  /api/v1/router/odoo/manage-activity:
    post:
      tags:
        - Universal Router
      summary: Router Gateway - Điều phối thêm / sửa / xóa Hoạt động (Ghi chú giao việc)
      description: >
        Thực thi tạo mới, chỉnh sửa hoặc hủy hoạt động mail.activity trên máy chủ Odoo đích được chỉ định bởi Router.
      operationId: routerManageActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouterActivityPayload'
      responses:
        '200':
          description: Thao tác hoạt động thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  action:
                    type: string
                    example: "update"
                  activitySummary:
                    type: string
                    nullable: true
                    example: "đã giao 4/9 xuat nhap PTQuyen"
                  target:
                    type: string
                    example: "odoo_test"
                  target_name:
                    type: string
                    example: "Odoo ERP (Môi trường Test)"

# ==============================================================================
# COMPONENTS SCHEMAS (CÁC ĐỊNH DẠNG DỮ LIỆU CHUẨN)
# ==============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    OrderUpdateLineItem:
      type: object
      required:
        - product_id
        - product_uom_qty
        - price_unit
      properties:
        id:
          type: integer
          nullable: true
          description: ID dòng sản phẩm Odoo (sale.order.line). Bỏ trống nếu là dòng thêm mới.
          example: 12580
        product_id:
          type: integer
          description: ID biến thể sản phẩm Odoo (product.product)
          example: 1042
        product_uom_qty:
          type: number
          minimum: 0
          description: Số lượng sản phẩm
          example: 5
        price_unit:
          type: number
          minimum: 0
          description: Đơn giá sản phẩm (VND)
          example: 25000
        discount:
          type: number
          minimum: 0
          maximum: 100
          default: 0
          description: Tỷ lệ chiết khấu theo dòng (%)
          example: 10
        name:
          type: string
          nullable: true
          description: Tên / mô tả dòng sản phẩm
          example: "Pate Thịt Cừu Cho Chó 100g"
        _deleted:
          type: boolean
          default: false
          description: Đánh dấu true nếu muốn xóa dòng sản phẩm này khỏi đơn hàng
          example: false
...
    RouterUpdateOrderPayload:
      type: object
      required:
        - odoo_order_id
        - lines
      properties:
        odoo_order_id:
          type: integer
          description: ID đơn hàng Odoo (sale.order)
          example: 2374
        lines:
          type: array
          description: Danh sách dòng sản phẩm cần cập nhật/thêm/xóa
          items:
            $ref: '#/components/schemas/OrderUpdateLineItem'
        client_order_ref:
          type: string
          nullable: true
          description: Ghi chú đơn hàng Odoo
          example: "Giao trong giờ hành chính"
        author_name:
          type: string
          nullable: true
          description: Tên nhân viên thao tác trên OCMS
          example: "Admin Quyen"
        author_id:
          type: integer
          nullable: true
          description: ID tài khoản Odoo của nhân viên thực hiện cập nhật
          example: 2
    RouterActivityPayload:
      type: object
      required:
        - odoo_order_id
        - action
      properties:
        odoo_order_id:
          type: integer
          description: ID đơn hàng Odoo (sale.order)
          example: 2374
        action:
          type: string
          enum: [create, update, delete]
          description: Hành động thao tác ghi chú giao việc
          example: "update"
        summary:
          type: string
          nullable: true
          description: Nội dung ghi chú giao việc (Hoạt động)
          example: "đã giao 4/9 xuat nhap PTQuyen"
        author_id:
          type: integer
          nullable: true
          description: ID tài khoản Odoo của người được giao việc
          example: 2

    # ── ĐƠN HÀNG ĐẦU VÀO (REQUEST) ──────────────────────────────────────────
...